+++ client/src/pages/List/List.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import styled from 'styled-components';
 import { useGlobalContext } from '../../Context'
-import ListRests from './ListRests';
-import ListMap from './ListMap';
+import ListRests from './ListRests/ListRests';
+import ListMap from './ListMap/ListMap';
 
 const List = () => {
   const [restId, setRestId] = useState(-1);
