+++ client/src/pages/List/ListRests/RestsItem/ItemInfo.js
