+++ client/src/pages/List/ListRests/RestsItem/ItemInfoFooter.js
@@ -2,6 +2,50 @@ import React from 'react';
 import styled from 'styled-components';
 import { BsStarFill } from 'react-icons/bs';
 import { flexAlign } from '../../../../Styles/theme';
+import { useGlobalContext } from '../../../../Context';
+
+const ItemInfoFooter = ({ mainInfo, subInfo, price, like }) => {
+  const { searchValue } = useGlobalContext();
+  const { maxGuest, bedroom, bed, bathroom } = mainInfo;
+  const { likeAvg, likeCount } = like;
+
+  return (
+    <InfoFooter>
+      <section className="info-wrapper">
+        <div>
+          <span>최대 인원 {maxGuest}명</span>
+          <span>∙</span>
+          <span>{bedroom === 0 ? '원룸' : `침실 ${bedroom}개`}</span>
+          <span>∙</span>
+          <span>침대 {bed === 0 ? '없음' : `${bed}개`}</span>
+          <span>∙</span>
+          <span>{bathroom === 0 ? '간이 욕실' : `욕실 ${bathroom}개`}</span>
+        </div>
+        <div>
+          {subInfo && subInfo.map((info, idx) => {
+            return (
+              <span>
+                {info}
+                {idx !== subInfo.length - 1 && <span>∙</span>}
+              </span>
+            )
+          })}
+        </div>
+      </section>
+      <section className="like-price-wrapper">
+        <div className="like">
+          <BsStarFill />
+          <strong>{likeAvg}</strong>
+          <span>({likeCount})</span>
+        </div>
+        <div className="price">
+          <p><strong>₩{price.toLocaleString()}</strong>/1박</p>
+          <span>총 요금: ₩{(price * searchValue.dateDiff).toLocaleString()}</span>
+        </div>
+      </section>
+    </InfoFooter>
+  )
+}
 
 const InfoFooter = styled.footer`
   display: flex;
@@ -43,47 +87,5 @@ const InfoFooter = styled.footer`
     }
   }
 `;
-
-const ItemInfoFooter = ({ mainInfo, subInfo, price, like }) => {
-  const { maxGuest, bedroom, bed, bathroom } = mainInfo;
-  const { likeAvg, likeCount } = like;
-
-  return (
-    <InfoFooter>
-      <section className="info-wrapper">
-        <div>
-          <span>최대 인원 {maxGuest}명</span>
-          <span>∙</span>
-          <span>{bedroom === 0 ? '원룸' : `침실 ${bedroom}개`}</span>
-          <span>∙</span>
-          <span>침대 {bed === 0 ? '없음' : `${bed}개`}</span>
-          <span>∙</span>
-          <span>{bathroom === 0 ? '간이 욕실' : `욕실 ${bathroom}개`}</span>
-        </div>
-        <div>
-          {subInfo && subInfo.map((info, idx) => {
-            return (
-              <span>
-                {info}
-                {idx !== subInfo.length - 1 && <span>∙</span>}
-              </span>
-            )
-          })}
-        </div>
-      </section>
-      <section className="like-price-wrapper">
-        <div className="like">
-          <BsStarFill />
-          <strong>{likeAvg}</strong>
-          <span>({likeCount})</span>
-        </div>
-        <div className="price">
-          <p><strong>₩{price.toLocaleString()}</strong>/1박</p>
-          <span>총 요금: ₩{(price * 2).toLocaleString()}</span>
-        </div>
-      </section>
-    </InfoFooter>
-  )
-}
 
 export default ItemInfoFooter;
